restructure classes into separate modules
